+++ ds3231/CHANGELOG.md
@@ -1,4 +1,7 @@
 # Change Log
 
+## [1.0] - 2018-07-23
+- update to Android Things 1.0
+
 ## [0.1] - 2017-12-13
 - initial version