+++ hpm/CHANGELOG.md
@@ -1,7 +1,11 @@
 # Change Log
 
-## [0.1] - 2017-11-21
-- initial version
+## [1.0] - 2018-07-23
+- update to Android Things 1.0
 
 ## [0.2] - 2017-12-04
 - initial version
+
+## [0.1] - 2017-11-21
+- initial version
+